requirements.txt
alembic==1.4.3
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
et-xmlfile==1.1.0
Flask==1.1.2
Flask-APScheduler==1.12.3
Flask-Caching==1.9.0
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
greenlet==1.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
openpyxl==3.0.9
Pillow==7.0.0
PyMySQL==0.10.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.3.20
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==1.0.1
zipp==3.8.0
